\name{cmsRaw}
\docType{data}
\alias{cmsRaw}
\title{An example cmsRaw object}
\description{
  This object contains parsed raw data for 4 samples in the 
  MTBLS2 dataset.
}
\usage{cmsRaw}
\format{A \code{CMSraw} object containing information on 4 samples
in the MTBLS2 dataset.}
\source{The \code{mtbls2} Bioconductor data package.}
\keyword{datasets}