\name{cmsRawExample}
\docType{data}
\alias{cmsRawExample}
\title{An example cmsRaw object}
\description{
  This object contains parsed raw data for 4 samples in the 
  MTBLS2 dataset.
}
\usage{cmsRawExample}
\value{
An object of class \code{CMSraw} containing parsed data for 4 samples 
in the MTBLS2 dataset.
}
\format{A \code{CMSraw} object containing information on 4 samples
in the MTBLS2 dataset.}
\source{The \code{mtbls2} Bioconductor data package.}
\keyword{datasets}